+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/social/KakaoUtil.kt
@@ -1,12 +1,11 @@
-package com.wafflestudio.toyproject.memoWithTags.user
+package com.wafflestudio.toyproject.memoWithTags.social
 
 import com.fasterxml.jackson.core.JsonProcessingException
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import com.wafflestudio.toyproject.memoWithTags.exception.OAuthRequestException
-import com.wafflestudio.toyproject.memoWithTags.social.dto.NaverOAuthToken
-import com.wafflestudio.toyproject.memoWithTags.social.dto.NaverProfile
-import com.wafflestudio.toyproject.memoWithTags.social.dto.NaverProfileResponse
+import com.wafflestudio.toyproject.memoWithTags.social.dto.KakaoOAuthToken
+import com.wafflestudio.toyproject.memoWithTags.social.dto.KakaoProfile
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpEntity
@@ -17,38 +16,35 @@ import org.springframework.stereotype.Component
 import org.springframework.util.LinkedMultiValueMap
 import org.springframework.util.MultiValueMap
 import org.springframework.web.client.RestTemplate
-import java.util.UUID
 
 @Component
-class NaverUtil(
-    @Value("\${naver.auth.client-id}")
-    private val naverClientId: String,
-    @Value("\${naver.auth.client-secret}")
-    private val naverClientSecret: String
+class KakaoUtil(
+    @Value("\${kakao.auth.client}")
+    private val kakaoClient: String,
+    @Value("\${kakao.auth.redirect}")
+    private val kakaoRedirect: String
 ) {
     private val logger = LoggerFactory.getLogger(javaClass)
 
-    fun requestToken(accessCode: String): NaverOAuthToken {
+    fun requestToken(accessCode: String): KakaoOAuthToken {
         val restTemplate = RestTemplate()
         val headers = HttpHeaders()
         headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8")
 
-        val state = UUID.randomUUID().toString()
         val params: MultiValueMap<String, String> = LinkedMultiValueMap<String, String>().apply {
             add("grant_type", "authorization_code")
-            add("client_id", naverClientId)
-            add("client_secret", naverClientSecret)
+            add("client_id", kakaoClient)
+            add("redirect_url", kakaoRedirect)
             add("code", accessCode)
-            add("state", state)
         }
 
-        val naverTokenRequest = HttpEntity(params, headers)
-        logger.info("Token Request: $naverTokenRequest")
+        val kakaoTokenRequest = HttpEntity(params, headers)
+        logger.info("Token Request: $kakaoTokenRequest")
 
         val response: ResponseEntity<String> = restTemplate.exchange(
-            "https://nid.naver.com/oauth2.0/token",
+            "https://kauth.kakao.com/oauth/token",
             HttpMethod.POST,
-            naverTokenRequest,
+            kakaoTokenRequest,
             String::class.java
         )
         logger.info("Token Response: $response")
@@ -56,7 +52,7 @@ class NaverUtil(
         val objectMapper = ObjectMapper().registerKotlinModule()
 
         return try {
-            val oAuthToken = objectMapper.readValue(response.body, NaverOAuthToken::class.java)
+            val oAuthToken = objectMapper.readValue(response.body, KakaoOAuthToken::class.java)
             logger.info("oAuthToken: ${oAuthToken.access_token}")
             oAuthToken
         } catch (e: JsonProcessingException) {
@@ -65,29 +61,29 @@ class NaverUtil(
         }
     }
 
-    fun requestProfile(oAuthToken: NaverOAuthToken): NaverProfile {
+    fun requestProfile(oAuthToken: KakaoOAuthToken): KakaoProfile {
         val restTemplate = RestTemplate()
         val headers = HttpHeaders().apply {
             add("Content-type", "application/x-www-form-urlencoded;charset=utf-8")
             add("Authorization", "Bearer ${oAuthToken.access_token}")
         }
 
-        val naverProfileRequest: HttpEntity<MultiValueMap<String, String>> = HttpEntity(headers)
-        logger.info("Profile Request: $naverProfileRequest")
+        val kakaoProfileRequest: HttpEntity<MultiValueMap<String, String>> = HttpEntity(headers)
+        logger.info("Profile Request: $kakaoProfileRequest")
 
         val response: ResponseEntity<String> = restTemplate.exchange(
-            "https://openapi.naver.com/v1/nid/me",
+            "https://kapi.kakao.com/v2/user/me",
             HttpMethod.GET,
-            naverProfileRequest,
+            kakaoProfileRequest,
             String::class.java
         )
         logger.info("Profile Response: $response")
 
         val objectMapper = ObjectMapper().registerKotlinModule()
         return try {
-            val naverProfile = objectMapper.readValue(response.body, NaverProfileResponse::class.java)
-            logger.info("naver email: $naverProfile")
-            naverProfile.response
+            val kakaoProfile = objectMapper.readValue(response.body, KakaoProfile::class.java)
+            logger.info("kakao email: ${kakaoProfile.kakao_account.email}")
+            kakaoProfile
         } catch (e: JsonProcessingException) {
             logger.info("Profile processing error: ${e.message}")
             throw OAuthRequestException()
