+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/social/service/SocialLoginService.kt
@@ -8,10 +8,10 @@ import com.wafflestudio.toyproject.memoWithTags.social.dto.KakaoProfile
 import com.wafflestudio.toyproject.memoWithTags.social.dto.NaverOAuthToken
 import com.wafflestudio.toyproject.memoWithTags.social.dto.NaverProfile
 import com.wafflestudio.toyproject.memoWithTags.social.dto.SocialLoginResponse
-import com.wafflestudio.toyproject.memoWithTags.user.GoogleUtil
+import com.wafflestudio.toyproject.memoWithTags.social.GoogleUtil
 import com.wafflestudio.toyproject.memoWithTags.user.JwtUtil
-import com.wafflestudio.toyproject.memoWithTags.user.KakaoUtil
-import com.wafflestudio.toyproject.memoWithTags.user.NaverUtil
+import com.wafflestudio.toyproject.memoWithTags.social.KakaoUtil
+import com.wafflestudio.toyproject.memoWithTags.social.NaverUtil
 import com.wafflestudio.toyproject.memoWithTags.user.SocialType
 import com.wafflestudio.toyproject.memoWithTags.user.controller.User
 import org.slf4j.LoggerFactory
